AI and Building changes
The building has been change from using ono begin or end overlap to using a custom even that gets the overlapped actors on event tick, this fixes the issues of the building moving to fast and not trigging the overlap.

Secondly changes to the AI has been made so that it now used a crowd control compoment. Though is improve more changes need to be made to made it more accurate.